python/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "tw-ai"
version = "0.1.0"
description = "AI components for Triage Warden - LLM integration, agents, and RAG"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "Triage Warden Team" }
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Information Technology",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Security",
]
dependencies = [
    "openai>=1.0.0",
    "anthropic>=0.18.0",
    "pydantic>=2.0.0",
    "httpx>=0.25.0",
    "tenacity>=8.2.0",
    "tiktoken>=0.5.0",
    "structlog>=23.0.0",
    "chromadb>=0.4.0",
    "sentence-transformers>=2.2.0",
    "pyyaml>=6.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
    "mypy>=1.0.0",
    "ruff>=0.1.0",
    "black>=23.0.0",
    "testcontainers[postgres,redis]>=4.0.0",
    "psycopg2-binary>=2.9.0",
    "qdrant-client>=1.7.0",
    "respx>=0.21.0",
]
local = [
    "vllm>=0.3.0",
    "transformers>=4.36.0",
    "torch>=2.0.0",
]

[project.urls]
Homepage = "https://github.com/triage-warden/triage-warden"
Documentation = "https://triage-warden.github.io/docs"
Repository = "https://github.com/triage-warden/triage-warden"

[tool.hatch.build.targets.wheel]
packages = ["tw_ai"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.mypy]
python_version = "3.10"
strict = true
ignore_missing_imports = true

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.ruff.lint.per-file-ignores]
# Prompt files contain JSON examples that are naturally long
"tw_ai/agents/prompts/*.py" = ["E501"]
# Benchmark files contain test data with long descriptions
"tw_ai/benchmarks/datasets.py" = ["E501"]
"tw_ai/benchmarks/tasks.py" = ["E501"]
"tw_ai/benchmarks/metrics.py" = ["E501"]

[tool.black]
line-length = 100
target-version = ["py310"]
